[feature/proxy] allow for proxying
These are changes I've made to allow for the possibility of fronting the
web-asset-server with a proxy or alternate web server (e.g., Apache's
mod_proxy).

I need to run the web-asset-server on a non-privileged port, since I
don't want to run it as root.  But I also want it to be accessible to
the public Internet, and local firewall policy in my organization
doesn't allow public access to non-standard ports.

Since this will probably be public-facing, I'd like to possibly wrap SSL
around it in the near future (so, that's another reason for wanting to
front it with a proxy).